Add kuberntes operator install script

+++ install.d/09-basher.sh
@@ -2,7 +2,7 @@
 # vim: ts=2 sw=2 noet :
 set -e
 
-INSTALL_PATH=$(dirname $0)
+INSTALL_PATH=$(dirname "$0")
 . $INSTALL_PATH/.env.sh
 
 sudo apt install -y git
@@ -14,12 +14,12 @@ if [ ! -d $BASHER_PATH ]; then
     git clone https://github.com/basherpm/basher.git $BASHER_PATH
 fi
 
-if [ ! -d ~/$PROFILE_DIR ]; then
-    rm -rf ~/$PROFILE_DIR
-    mkdir -p ~/$PROFILE_DIR
+if [ ! -d "${HOME:?}/$PROFILE_DIR" ]; then
+    rm -rf "${HOME:?}/$PROFILE_DIR"
+    mkdir -p "${HOME:?}/$PROFILE_DIR"
 fi
 
-cat <<-DONE > $BASHER_INIT
+cat <<-DONE > "$BASHER_INIT"
 export PATH="$HOME/.basher/bin:$PATH"
 eval "\$(basher init - bash)"
 DONE
